refactor : implement chart title with options (#3)
- type.chart.title.ts 부분 수정
