+++ src/type/index.ts
@@ -5,4 +5,4 @@ export {BackgroundStyle} from './type.background.style'
 export {DataFlow, GraphData} from './type.data.flow'
 export {TooltipStyle} from './type.tooltip.style'
 export {EventFlow} from './type.event.flow'
-export {ChartTitle, TitleType} from "./type.chart.title"
+export {ChartTitle, TitleType, TitleOptions} from "./type.chart.title"
