+++ src/type/type.chart.title.ts
@@ -2,7 +2,17 @@
 export interface ChartTitle{
     type : TitleType// 제목의 타입 설정
     content : string, // 제목의 내용
-    options?: any,
+    options : TitleOptions,
+}
+
+export interface TitleOptions{
+    font?:string,
+    fontSize?:number,
+    fontWeight?:string,
+    color?:string,
+    backgroundColor?:string,
+    shadow?:string,
+    link?:string,
 }
